Allow render to callback with POJOs
In certain scenarios, our "component" may not actually return a React
component, by design.  In this case, we can provide a custom render
function that will process the return value, and return a plain old
JavaScript object (POJO) as it sees fit.  When using the provided render
functions, previous functionality is maintained.  That means that if the
built-in render functions are used, the component is still expected to
only return a React component.